Updated: add a contrain to block user who's token does not have any admin abilities

+++ app/Http/Controllers/CtfCategoryController.php
@@ -23,6 +23,10 @@ class CtfCategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if(!$request->user()->tokenCan('admin:*')){
+            abort(403, 'Unauthorized. You do not have permission.');
+        }
+
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:ctf_categories,name',
             'color' => 'nullable|string|max:7',
@@ -40,8 +44,9 @@ class CtfCategoryController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(CtfCategory $ctfCategory)
+    public function show(string $id)
     {
+        $ctfCategory = CtfCategory::findOrFail($id);
         return response()->json([
             'success' => true,
             'data' => $ctfCategory->loadCount('ctfChallenges'),
@@ -51,8 +56,14 @@ class CtfCategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, CtfCategory $ctfCategory)
+    public function update(Request $request, string $id)
     {
+        if(!$request->user()->tokenCan('admin:*')){
+            abort(403, 'Unauthorized. You do not have permission.');
+        }
+
+        $ctfCategory = CtfCategory::findOrFail($id);
+
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:ctf_categories,name,' . $ctfCategory->id,
             'color' => 'nullable|string|max:7',
@@ -70,8 +81,14 @@ class CtfCategoryController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(CtfCategory $ctfCategoty)
+    public function destroy(string $id)
     {
+        if(!request()->user()->tokenCan('admin:*')){
+            abort(403, 'Unauthorized. You do not have permission.');
+        }
+
+        $ctfCategoty = CtfCategory::findOrFail($id);
+
         $ctfCategoty->delete();
 
         return response()->json(null, 204);
